+++ microscope/_wrappers/BMC.py
@@ -24,12 +24,14 @@ import ctypes
 import os
 from ctypes import c_char, c_char_p, c_double, c_int, c_uint, c_uint32
 
+import microscope._utils
+
 
 if os.name == "nt":  # is windows
     # Not actually tested yet
-    SDK = ctypes.WinDLL("BMC2")
+    SDK = microscope._utils.library_loader("BMC2", ctypes.WinDLL)
 else:
-    SDK = ctypes.CDLL("libBMC.so.3")
+    SDK = microscope._utils.library_loader("libBMC.so.3")
 
 
 # Definitions from BMCDefs.h