+++ microscope/_wrappers/asdk.py
@@ -24,13 +24,14 @@ import ctypes
 import os
 from ctypes import c_char_p, c_double, c_int, c_size_t, c_uint32
 
+import microscope._utils
+
 
 if os.name == "nt":  # is windows
     _libname = "ASDK"
 else:
     _libname = "libasdk.so"  # Not actually tested yet
-
-SDK = ctypes.CDLL(_libname)
+SDK = microscope._utils.library_loader(_libname)
 
 
 class DM(ctypes.Structure):