+++ microscope/stages/linkam.py
@@ -48,6 +48,7 @@ from ctypes import POINTER, byref
 from enum import Enum, IntEnum
 
 import microscope
+import microscope._utils
 import microscope.abc
 
 
@@ -975,9 +976,10 @@ class _LinkamBase(microscope.abc.FloatingDeviceMixin, microscope.abc.Device):
     def init_sdk():
         """Initialise the SDK and set up event callbacks"""
         if os.name == "nt":  # is windows
-            __class__._lib = ctypes.CDLL("LinkamSDK.dll")
+            _libname = "LinkamSDK.dll"
         else:  # assuming Linux.  Not tested.
-            __class__._lib = ctypes.CDLL("libLinkamSDK.so")
+            _libname = "libLinkamSDK.so"
+        __class__._lib = microscope._utils.library_loader(_libname)
         _lib = __class__._lib
         """Initialise the SDK, and create and set the callbacks."""
         # Omit conditional pending a fix for ctypes issues when optimisations in use.